requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
feedparser==6.0.10
beautifulsoup4==4.12.2
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pyyaml==6.0.1